Normalized the project infrastructure
