+++ packages/browserslist-config/src/index.ts
@@ -5,4 +5,7 @@ const config: string[] = [
     "not dead"
 ];
 
-export default config;
+// Using TypeScript "export" keyword until browserlist support ESM.
+// Otherwise we must deal with a weird CommonJS output from esbuild which is not worth it.
+// For more info, see: https://github.com/evanw/esbuild/issues/1079
+export = config;
