+++ packages/browserslist-config/tsup.config.ts
@@ -2,15 +2,11 @@ import { defineConfig } from "tsup";
 
 export default defineConfig({
     clean: true,
+    splitting: false,
+    treeshake: true,
     entry: ["./src"],
     outDir: "./dist",
     format: ["cjs"],
     target: "esnext",
-    platform: "node",
-    // View https://github.com/egoist/tsup/issues/572#issuecomment-1060599574.
-    esbuildOptions: options => {
-        options.footer = {
-            js: "module.exports = module.exports.default;"
-        };
-    }
+    platform: "node"
 });
