+++ packages/eslint-plugin/package.json
@@ -17,8 +17,12 @@
         "directory": "packages/eslint-plugin"
     },
     "license": "Apache-2.0",
-    "main": "./dist/index.js",
-    "types": "./dist/index.d.ts",
+    "exports": {
+        ".": {
+            "require": "./dist/index.js",
+            "types": "./dist/index.d.ts"
+        }
+    },
     "files": [
         "docs",
         "dist",
@@ -45,7 +49,8 @@
         "@types/estree": "1.0.0",
         "@typescript-eslint/parser": "5.58.0",
         "eslint": "8.38.0",
-        "tsup": "6.7.0"
+        "tsup": "6.7.0",
+        "typescript": "5.0.4"
     },
     "peerDependencies": {
         "@typescript-eslint/parser": "*",
