+++ packages/eslint-plugin/src/config/by-project-type/react-library.ts
@@ -1,0 +1,34 @@
+import { mdxFiles, sourceFiles } from "../../utils/patterns";
+
+import type { Linter } from "eslint";
+
+const config: Linter.Config = {
+    overrides: [
+        {
+            files: sourceFiles,
+            plugins: ["@workleap"],
+            extends: [
+                "plugin:@workleap/core",
+                "plugin:@workleap/typescript",
+                "plugin:@workleap/react",
+                "plugin:@workleap/jest",
+                "plugin:@workleap/testing-library",
+                "plugin:@workleap/storybook"
+            ],
+            rules: {
+                // Custom WorkLeap rules
+                "@workleap/strict-css-modules-names": "warn"
+            }
+        },
+        {
+            files: mdxFiles,
+            plugins: ["@workleap"],
+            extends: ["plugin:@workleap/mdx"]
+        }
+    ]
+};
+
+// Using TypeScript "export" keyword until ESLint support ESM.
+// Otherwise we must deal with a weird CommonJS output from esbuild which is not worth it.
+// For more info, see: https://github.com/evanw/esbuild/issues/1079
+export = config;
