+++ packages/eslint-plugin/src/config/by-project-type/typescript-library.ts
@@ -1,5 +1,6 @@
+import { mdxFiles, sourceFiles } from "../../utils/patterns";
+
 import type { Linter } from "eslint";
-import { sourceFiles, mdxFiles } from "../../utils/patterns";
 
 const config: Linter.Config = {
     overrides: [
@@ -9,15 +10,9 @@ const config: Linter.Config = {
             extends: [
                 "plugin:@workleap/core",
                 "plugin:@workleap/typescript",
-                "plugin:@workleap/react",
                 "plugin:@workleap/jest",
-                "plugin:@workleap/testing-library",
-                "plugin:@workleap/storybook"
-            ],
-            rules: {
-                // Custom WorkLeap rules
-                "@workleap/strict-css-modules-names": "warn"
-            }
+                "plugin:@workleap/testing-library"
+            ]
         },
         {
             files: mdxFiles,
@@ -27,4 +22,7 @@ const config: Linter.Config = {
     ]
 };
 
+// Using TypeScript "export" keyword until ESLint support ESM.
+// Otherwise we must deal with a weird CommonJS output from esbuild which is not worth it.
+// For more info, see: https://github.com/evanw/esbuild/issues/1079
 export = config;
