+++ packages/eslint-plugin/src/config/core.ts
@@ -146,4 +146,7 @@ const config: Linter.Config = {
     }
 };
 
+// Using TypeScript "export" keyword until ESLint support ESM.
+// Otherwise we must deal with a weird CommonJS output from esbuild which is not worth it.
+// For more info, see: https://github.com/evanw/esbuild/issues/1079
 export = config;
