+++ packages/eslint-plugin/src/config/storybook.ts
@@ -1,6 +1,7 @@
 // https://github.com/storybookjs/eslint-plugin-storybook/blob/main/lib/configs/recommended.ts
 
-import { storybookFiles, mainStorybookFiles } from "../utils/patterns";
+import { mainStorybookFiles, storybookFiles } from "../utils/patterns";
+
 import type { Linter } from "eslint";
 
 const config: Linter.Config = {
@@ -23,4 +24,7 @@ const config: Linter.Config = {
     ]
 };
 
+// Using TypeScript "export" keyword until ESLint support ESM.
+// Otherwise we must deal with a weird CommonJS output from esbuild which is not worth it.
+// For more info, see: https://github.com/evanw/esbuild/issues/1079
 export = config;
