+++ packages/eslint-plugin/src/config/testing-library.ts
@@ -1,5 +1,6 @@
+import { reactTestFiles, testFiles } from "../utils/patterns";
+
 import type { Linter } from "eslint";
-import { testFiles, reactTestFiles } from "../utils/patterns";
 
 const config: Linter.Config = {
     overrides: [
@@ -15,4 +16,7 @@ const config: Linter.Config = {
     ]
 };
 
+// Using TypeScript "export" keyword until ESLint support ESM.
+// Otherwise we must deal with a weird CommonJS output from esbuild which is not worth it.
+// For more info, see: https://github.com/evanw/esbuild/issues/1079
 export = config;
