+++ packages/eslint-plugin/src/utils/patterns.ts
