+++ packages/eslint-plugin/tests/export.test.ts
@@ -1,10 +1,10 @@
 import fs from "fs";
 import path from "path";
-import plugin from "../lib";
+import plugin from "../src";
 
-const rules = fs.readdirSync(path.resolve(__dirname, "../lib/rules")).map(x => path.parse(x).name);
-const configsByUseCase = fs.readdirSync(path.resolve(__dirname, "../lib/config")).map(x => path.parse(x)).filter(x => x.ext).map(x => x.name);
-const configsByProjectType = fs.readdirSync(path.resolve(__dirname, "../lib/config/by-project-type")).map(x => path.parse(x).name);
+const rules = fs.readdirSync(path.resolve(__dirname, "../src/rules")).map(x => path.parse(x).name);
+const configsByUseCase = fs.readdirSync(path.resolve(__dirname, "../src/config")).map(x => path.parse(x)).filter(x => x.ext).map(x => x.name);
+const configsByProjectType = fs.readdirSync(path.resolve(__dirname, "../src/config/by-project-type")).map(x => path.parse(x).name);
 const configs = [...configsByUseCase, ...configsByProjectType];
 
 const exportedConfigs = Object.keys(plugin.configs ?? {});
