+++ packages/postcss-plugin/package.json
@@ -15,8 +15,12 @@
         "url": "git+https://github.com/workleap/wl-web-configs.git",
         "directory": "packages/postcss-plugin"
     },
-    "main": "./dist/index.js",
-    "types": "./dist/index.d.ts",
+    "exports": {
+        ".": {
+            "require": "./dist/index.js",
+            "types": "./dist/index.d.ts"
+        }
+    },
     "files": [
         "dist",
         "CHANGELOG.md",
@@ -29,9 +33,11 @@
         "postcss-preset-env": "8.3.1"
     },
     "devDependencies": {
+        "@workleap/eslint-plugin": "workspace:*",
         "@workleap/typescript-configs": "workspace:*",
         "postcss": "8.4.21",
-        "tsup": "6.7.0"
+        "tsup": "6.7.0",
+        "typescript": "5.0.4"
     },
     "peerDependencies": {
         "postcss": ">=8.4.6"
