+++ packages/postcss-plugin/src/index.ts
@@ -23,4 +23,7 @@ const plugin: PluginCreator<typeof DefaultOptions> = ({ presetEnvOptions } = Def
 
 plugin.postcss = true;
 
+// Using TypeScript "export" keyword until PostCSS support ESM.
+// Otherwise we must deal with a weird CommonJS output from esbuild which is not worth it.
+// For more info, see: https://github.com/evanw/esbuild/issues/1079
 export = plugin;
