+++ packages/postcss-plugin/tsconfig.json
@@ -1,7 +1,4 @@
 {
     "extends": "@workleap/typescript-configs/library.json",
-    "compilerOptions": {
-        "module": "CommonJS"
-    },
     "exclude": ["dist", "node_modules"]
 }
