+++ packages/stylelint-config/.stylelintignore
@@ -1,0 +1,1 @@
+*
