+++ packages/stylelint-config/README.md
@@ -1,4 +1,5 @@
-# `@workleap/stylelint-config`
+# @workleap/stylelint-config
+
 Shareable [stylelint](https://stylelint.io/) configuration for Workleap.
 
 [![License](https://img.shields.io/badge/License-Apache_2.0-blue.svg)](../../LICENSE)
@@ -9,6 +10,7 @@ Shareable [stylelint](https://stylelint.io/) configuration for Workleap.
 Install the package.
 
 **With pnpm**
+
 ```shell
 pnpm add -D @workleap/stylelint-config
 ```
