+++ packages/stylelint-config/src/index.ts
@@ -73,4 +73,7 @@ const config: Config = {
     }
 };
 
+// Using TypeScript "export" keyword until StyleLint support ESM.
+// Otherwise we must deal with a weird CommonJS output from esbuild which is not worth it.
+// For more info, see: https://github.com/evanw/esbuild/issues/1079
 export = config;
