+++ packages/tsup-configs/package.json
@@ -30,8 +30,10 @@
         "build": "tsup"
     },
     "devDependencies": {
+        "@workleap/eslint-plugin": "workspace:*",
         "@workleap/typescript-configs": "workspace:*",
-        "tsup": "6.7.0"
+        "tsup": "6.7.0",
+        "typescript": "5.0.4"
     },
     "peerDependencies": {
         "tsup": "*"
