+++ packages/tsup-configs/src/build.ts
@@ -3,9 +3,10 @@ import { type DefineConfigOptions, mergeConfigs } from "./mergeConfigs.ts";
 export function defineBuildConfig(config?: DefineConfigOptions) {
     return mergeConfigs(config, {
         clean: true,
-        treeshake: true,
-        minify: true,
         dts: true,
+        minify: true,
+        splitting: false,
+        treeshake: true,
         entry: ["./src"],
         outDir: "./dist",
         format: ["esm"],
