+++ packages/tsup-configs/tsup.config.ts
@@ -3,10 +3,11 @@ import { defineConfig } from "tsup";
 export default defineConfig({
     clean: true,
     dts: true,
-    minify: true,
+    splitting: false,
     treeshake: true,
     entry: ["src"],
     outDir: "./dist",
     format: ["esm"],
-    target: "esnext"
+    target: "esnext",
+    platform: "node"
 });
